Changed the row with the survey on the end date in the database in the queries table

+++ src/main/java/com/gazprom/InforamtionSystem/repository/RoleRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 @Repository
 public interface RoleRepository extends JpaRepository<Role, Long> {
     Optional<Role> findByRole(String role);
-    Optional<Role> findById(Long id);
+    Optional<Role> findById(int id);
 
 }
